rtl: sc-cm3-ss: Release the low fixing of CDBGPWRUPACK
The Cortex-M3 debug port has a mechanism to control the power supply
of the system.

When the debug controller sets CDBGPWRUPREQ High, the power controller
powers on the debug port or system. The power controller returns the
signal to the CDVGPWRUPACK. After checking this signal, the debug port
can start debugging.

sc-obc-a1-fpga does not control the power supply by CDBGPWRUPREQ. So
connect CDBGPWRUPREQ and CDBGPWRUPACK.