implement: synthesis: Delete reset_timing command
In the initial logic synthesis, the reset_timing command was inserted
for the purpose of deleting the timing constraint added in the sub
module hierarchy.

In the design flow of this design, the timing constraint added in the
logic synthesis of sc_obc_a1_fpga is inherited. Use the timing_clock*
command to confirm that the set timing is appropriate.