rtl: Add BOOT_RSTB signal
BOOR_RSTB is a reset signal that is asserted only once when the FPGA
boot. It is asserted until the reference clock is selected.  It is the
only reset signal used by the logic that prepares the system.

Other reset signals output by the system controller will not be
released until the system is ready.