rtl: Replaced Flip Flop for latching ITCMEN
This Flip Flop drives the clock signal at the rising edge of the
reset. This is because we want to latch the signal only once when the
reset is released. However, in the conventional circuit, the clock and
reset are racing. To avoid this, we have implemented this Flip Flop
without asynchronous reset. And then, the initial value of the
Flip Flop was given, and the value was fixed after the configuration.