main_axi_ss: Rename the signal that connects to the AXI crossbar
All signal names with "axi" in this module are connected to the AXI
crossbar. So, the prefix "w_crsbar_" does not contribute to the ease
of code reading. I will shorten the signal names to simplify these.

Also, consider the axi namespace and change to "axis" and "axim"
instead of the "s_axi" and "m_axi" prefixes. This means that there
are slave and master in the axi standard.