tb: Add task to skip SRAM initialization
SRAM initialization executed after power-on requires 87.4 ms in the
worst case. The initialization sequence does not depend on the
operating mode of SC OBC FPGA and is always executed. The HRMEM IP
core does not have the function to skip the SRAM initialization
sequence when it is not needed. So we need to have it in the test
bench. This is necessary to reduce the simulation time (reduce the
development time).