sc-cm3-ss: Update AHB Multiplexer
The original logic had a problem when HREADY went low.

In addition, I-Code and D-Code AHB were supposed to operate
exclusively. But, when HREADY is Low, there are cases where it
seems that requests are being received at the same time.

So, we changed the circuit to receive and process requests from I-Code
and D-Code in a double buffer.