sc-hrmem: ahb_sram_ctrl: Fixed control when transfer type is Busy
When Busy was inserted during AHB Read Burst transfer, the data phase
transmission stopped at the same time as waiting for the next address
phase, and transactions were stuck.

For this reason, at the time of Busy, the transmission of the data
phase was completed, and only the Wait of the next address phase was
corrected.

Also, it was corrected because multiple RAM writes were illegally
performed even during Busy of Write Burst.