sc-hrmem: Fixed control when no Interval between AHB Burst Read to Prefetch
When Prefetch Read was started without an interval from the completion
of AHB Burst Read, the Prefetch access to RAM was rewritten to the
invalid RAM access by the look-ahead access of Burst Read.

For this reason, in such a case, the RAM is accessed with priority
given to Prefetch access.