sc-hrmem: ahb_sram_ctrl: Fixed invalid data mask period after AHB Burst Read
When the next AHB Burst Read was started without an interval from the
completion of AHB Read, the read data due to the invalid read-ahead
access of the burst read completed earlier was erroneously detected as
the read data of the later Read.

For this reason, a condition was added to extend the masked period of
invalid read data after the completion of AHB Burst Read and to
exclude it from masking correct access.