sc-can: Changed read data control at message resend start
priority inversion As a preparation for avoidance measures, switch the
message to be sent to the latest data even when the frame is
retransmitted after arbitration or error occurs.

For this reason, the reading from the TX FIFO and TX High Priority
Buffer is executed even at the start of retransmission, and the
increment timing of the read pointer is changed from the reading time
to the completion of normal transmission.

If the TX FIFO is reset during transmission, Not EMPTY is immediately
negated, but the TX FIFO UDF is not generated when the transmission is
completed.