sc-can: Add TX Message Priority Control module
A control module for the function of transmitting the transmitted
messages stored in the TX FIFO in order from the highest priority
message has been added.

In this commit, only the created new module is added, and the
implementation in the CAN module is done in the next commit.

The functional outline of the TX message priority control module is as
follows:

  - TX FIFO uses Memory Module(sc_can_mem) to control write and read
    pointers from outside the Memory Module.

  - The priority order management of each TX message and the pointer
    control and status management of the TX FIFO are managed by the TX
    Priority Controller Module(sc_can_tx_prio_ctl).

  - When a TX message is stored in the TXFIFO, a preferred search for
    the TX message is performed.

  - When the message is ready to be sent, the highest priority TX
    message at that time is read from the TXFIFO and sent in sequence
    to the CAN bus.

  - However, it is the same as the conventional specification that the
    transmission of the stored message of TX High Priority Buffer has
    the highest priority.