sc_fifo_async: Add data counter sync clock selection function
Data counter of the sc_fifo_async module was synchronized with the
clock on the write side of the data, it was modified so that the
synchronization clock of the data counter could be selected to
synchronize with the clock on the write side or the read side.

Since Default is set to synchronize with the clock on the Write side,
there is no effect on the circuit at the time of this commit. The
setting of the synchronization clock selection in the upper layer is
done in the next commit.