sc-can: can_msg_strg: RX FIFO data counter changed to Read clock synchronization
Since the RX FIFO data counter was synchronized with the clock on the
Write (CAN Core) side, a case was found in which the counter decrement
cannot operate normally when the RX FIFO data is read from the Read
(Bus) side at high speed and continuously.

For this reason, the data counter was modified to clock
synchronization on the Read (Bus) side.