sc-can: Optimized TXF_FULL, RXF_FULL, INT_RXFVAL signal generation
These signals are generated by decoding the data counter value of the
FIFO, but once the data counter signal converted to CAN Clock
synchronization is decoded by the Bit Stream Processor, the decoded
signal is converted to Bus Clock again and output to Register.

For this reason, it was changed to output the signal decoded by
Massage Storage directly to the register without performing Clock
conversion.