rtl: ahb_sram_ctrl: Fixed AHB transaction change detection when BUSY to NONSEQ transition
At the time of Burst transfer, it was assumed that SEQ was always
inserted after the transfer type (HTRANS) was BUSY. When the NONSEQ
was inserted after BUSY and the next AHB transaction was started, the
internal state machine could not detect the transaction switch and the
memory access could not work properly.

For this reason, when NONSEQ is detected after BUSY, HREADY is negated
for 1 cycle, fixed to detect the start of the next transaction.