sc-hrmem: ahb_sram_ctrl: Fixed read data storage F/F initialization conditions
The initialization condition of the F/F that temporarily stores the
read data read ahead during the BUSY of the AHB Read Burst transaction
was set at the Idle of the internal state machine. As a result, when
Read Burst transactions occur continuously without any interval, the
internal state machine does not transition to Idle, so initialization
is not performed and invalid data is sent in the subsequent Read
Burst.

For this reason, the initialization condition was modified at the end
of the Read transaction.