sc-hrmem: ahb_sram_ctrl: Fixed read-ahead control stop condition at AHB Read Burst
When there was only one free space left read data storage F/F, reading
to memory was stopped, but the Busy period between beats of the Burst
Read transaction became longer and reading from the read data storage
F/F Overflow occurred when was slowed down.

For this reason, the stop condition for read-ahead control of read
data was changed when there were only two free space left read data
storage F/F, and it was corrected so that overflow would not occur.