sc-hrmem: ahb_sram_ctrl: Fixed read-ahead data transmission timing during AHB Read Burst
The read timing of the data read ahead and stored in the F/F during
the Wait of the AHB Read Burst transaction has shifted to the address
phase of the target beat. As a result, we found a case where correct
read data could not be sent in the data phase.

For this reason, mask control was added so that the F/F data is not
read at an incorrect timing.