sc-hrmem: ahb_sram_ctrl: Added mask exclusion of read data at the end of AHB Read Burst
When the AHB Read Burst transitions to IDLE or NONSEQ after BUSY
during a transaction and the next transaction is interrupted and only
1 Beat is transferred, the timing is the same as the address phase of
the next Read transaction, and the data phase of the previous
transaction The lead data transmission was masked.

For this reason, it was corrected so that it would not be masked.