sc-hrmem: ahb_sram_ctrl: Added mask for invalid memory read data at end of AHB Read Burst
When AXI Write is started at the same time as the end of the AHB Read
Burst transaction, it was found that the memory address read that
becomes unnecessary due to the end of the indefinite length Burst
control conflicts with the memory write by AXI Write.

For this reason, mask control has been added so that read control that
is sent to memory at the same time as the end of the AHB Read Burst
transaction is not performed.