sc-hrmem: ahb_sram_ctrl: Added mask for rogue data output during AHB Burst
If a Data Phase is received at BUSY during an AHB Burst transaction,
at the next NON SEQ or Address Phase at SEQ, In the case of Write, the
same data was written to the memory at the same address as the
previous time, and in the case of Read, unnecessary data was output to
the AHB Bus.

For this reason, when Data Phase processing is performed by Busy, mask
control is added so that unnecessary data processing is not performed
in the next Address Phase.