sc-cm3-ss: ahb_mux: Added control to stop burst transfer of memory bus after IDLE on CPU bus
When the memory bus was in the BUSY state, burst transfer of the
memory bus did not stop immediately even if the CPU bus was in the
IDLE state. At this time, the memory bus extends the BUSY state
instead of inserting an IDLE. This behavior is not a violation of the
AHB specifications. However, it is unfriendly as a bus control.