axi_slave: Fixed access when Offset address is non-zero
It was designed as an operation to shift the position of data on the
AXI Slave side according to the value of the address offset supplied
from the AXI Master when receiving AXI access other than ADDR[1:0]=0.

However, to be correct, the data position is already shifted from the
AXI Master side, so it has been corrected so that the data position is
not shifted on the Slave side.