sc-hrmem: hrmem_reg: Symbol name "AXIxx" & "AHBxx" unified to "BUSRDxx"
The modification target is only the signal name, Define name, and Port
name of the hrmem_reg module.
If you change everything except the hrmem_reg module, the scale of the
fix is very large and there is a risk of bugs being mixed in, and
since it is a problem only with the signal name inside the IP core and
there is no actual harm as it is, so in this commit will not fix it.