system-monitor: Change the read value of the WSR field of Watchdog Control Register
The value read from the WSR field of the Watchdog Control Register was
0x0000. The WSR field of the Watchdog Control Register read 0x0000.
This read value is changed to read the next write value for reloading
the Software Watchdog Timer. For example, if the next value to write
is 0x5A5A, the read value of this register is 0x5A5A. This can be used
for synchronization of software and Software Watchdog Timer.