system-monitor: Add Watchdog Mode Register
The Watchdog Mode Register sets the behavior when the Watchdog Timer
expires.

When this register is clear to 0, the FPGA requests a power cycle to
the TRCH by stopping toggle the FPGA_WATCHDOG signal. When this
register is set to 1, the FPGA restarts the system by resetting to the
FPGA logic.

This function was accidentally removed in commit "ff68d148".