system-monitor: Add SEM Controller module
SEM Controller is a Wrapper module for implementing Xilinx's SEM Core.
The role of this module is to connect the SEM Core to the FPGA's ICAP2
and FRAME_ECCE2 primitives.

SEM Core control is handled asynchronously by System Monitor
registers. The external signals of this module are structured to
facilitate asynchronous passing.