sc-hrmem: Changed SRAM I/F signal to sequential circuit output
The signal output of the SRAM interface was configured with a
combinational circuit in order to suppress the latency, but the
configuration was such that when the output delay difference of the
internal signal became large, a glitch would occur. For this reason,
the signal output of the SRAM I/F was changed to a sequential circuit.

Along with this, the latency when reading data from SRAM increases by
1 Cycle, so we adjusted the related timing accordingly.