sc-hrmem: added latency switching mode for SRAM read access
If the SRAM timing specification (tACE=10ns[max]) from the assertion
of Chip Enable to when the SRAM output data is valid can be satisfied
sufficiently with the time of one cycle width of the system clock,　
added a function to advance the timing of latching SRAM output data by
1 cycle and shorten the latency of read access to SRAM by 1 cycle.

Depending on the state of the added mode selection signal
(RD_LTCY_MODE), the read latency is switched as follows.
- RD_LTCY_MODE = 0: 4Cycles
- RD_LTCY_MODE = 1: 5Cycles

The input of this mode selection signal is connected to the high-order
1 bit of the CLK mode selection signal that reflects the set value
after the operating frequency setting is propagated to the clock
generation circuit.