rtl: sc-obc-a1-fpga: Chang clock pin of configuration flash from CFG_MEM_SCL to CCLK
On the OBC module board, CCLK and CFG_MEM_SCK are
connected. Traditionally, after configuration, the CCLK pin was in a
Hi-Z state and did not conflict with CFG_MEM_SCK.

However, after implementing the SEM Controller, the CCLK pin was stuck
low. And it conflicted with the operation of CFG_MEM_SCK.

We tried to work around this problem with USRCCLKTS in the STARTUPE2
primitive, but that didn't work. So we changed the clock pin of the
configuration flash memory.