uartlite: Added reset transmit Prescale counter when Baudrate changed
When the Baudrate is changed by register setting, if the transmission
Prescale counter value is larger than the changed Prescale value, the
transmission cannot be started unless the transmission Prescale
counter rolls over once.

For this reason, the transmission Prescale counter is reset when the
Baudrate setting value is changed.