sc-obc-core: Implement Read Latency Mode signal
The Read Latency Mode signal sets the HRMEM access mode.

Read Latency Mode is 0:
The SRAM read timing is 2 cycles.

Read Latency Mode is 1:
The SRAM read timing is 1 cycle.

When the system clock is 96MHz, reading SRAM data requires 2 cycles.
When the system clock is changed to 96MHz, the Read Latency Mode
should be changed before the clock switches. And when the system clock
is other than 96MHz, the Read Latency Mode needs to be changed after
the clock is switched.