lpahb-ss: Corrected I2C waveform parameters to 400 kbps (System Clock 48 MHz)
Its setting was 400 Kbps (System Clock 96 MHz). This is a bug in a
previous commit.

External I2C does not know what kind of device is connected. So the
initial value of I2C is not set to high speed mode. These settings can
be changed at runtime by software.